Every time an IRQ occurs the kernel reads from a GICv2 distributor configuration register to determine whether the interrupt is edge-triggered or level-triggered. If it is edge-triggered, the pending status of the interrupt must be cleared. The read from the distributor was observed to be expensive and so this pull request implements a cache for the interrupt configuration registers.

The following seL4Bench benchmarks demonstrate the performance improvement of this change using unicore MCS as measured from userspace (100 runs of IRQUser "With context switch" benchmark):

| Platform | cycles before: mean (stddev) | cycles after: mean (stddev) | diff: absolute (%) |
|---|---|---|---|
| OdroidC2 | 1132 (7) | 904 (7) | 228 (20%) |
| OdroidC4 | 1111 (7) | 889 (6) | 222 (20%) |
| TX1 | 1019 (7) | 813 (3) | 206 (20%) |