@Ivan-Velickovic 2023-09-25
It should be noted that GICv3 seems to not suffer from this issue.