@danshea00 2023-10-03
Thanks for the comments @Indanz! I have a few questions.

> At best another edge-triggered interrupt which came before the ack will get cleared with the current code and save one unnecessary interrupt by coalescing the two edge-triggered irqs (assuming that's wanted behaviour). However, if the interrupt comes after the ack, it will still be delivered as a new interrupt. The time window in the kernel is very small, while the time window between the signal and user space calling seL4_IRQHandler_Ack() is much bigger, so the current code isn't very effective.

I don’t see why the time window between `dist_pending_clr()` being called and the interrupt being reenabled by `seL4_IRQHandler_Ack()` is an issue. The interrupt is disabled prior to the `dist_pending_clr()` by `maskInterrupt()`. Can an interrupt become pending even if it is disabled?

> Clearing pending IRQs for edge-triggered ones would make much more sense when enabling an edge-triggered IRQ in maskInterrupt(), if that's what's wanted (and then your caching code makes sense too). 

Are you suggesting to move `dist_pending_clr()` within `seL4_IRQHandler_Ack()` (i.e. when `maskInterrupt()` reenables the interrupt)? This doesn’t make much sense to me, given that a lower priority pending interrupt would not be handled until the userspace calls `seL4_IRQHandler_Ack()` for the higher priority pending interrupt.

https://github.com/seL4/seL4/blob/61a8ccef44eeb08ecd20ad40296340c8319adee5/src/arch/arm/machine/gic_v2.c#L93-L96

Is this the “special code for level-triggered IRQs'' you are referring to?

It seems to me like for the current code, simply removing `is_irq_edge_triggered()` would make sense. The GICv2 docs indicate that this shouldn’t affect the pending status for level-triggered interrupts as the kernel never writes to `GICD_ISPENDR`.