@Indanz 2023-10-03
> Can an interrupt become pending even if it is disabled?

Yes. Disabling an IRQ has no effect on its state, it only stops interrupting the CPU.

> Are you suggesting to move dist_pending_clr() within seL4_IRQHandler_Ack() (i.e. when maskInterrupt() reenables the interrupt)?

_If_ special handling of edge-triggered interrupts is wanted, then yes, I propose it should be done when doing the ACK.

> This doesn’t make much sense to me, given that a lower priority pending interrupt would not be handled until the userspace calls seL4_IRQHandler_Ack() for the higher priority pending interrupt.

That's not how it works, priorities have nothing to do with it: The `dist_pending_clr()` clears the pending state of the currently active interrupt. It only has an effect if the IRQ state is active-and-pending, which happens for edge-triggered interrupts if another edge happens on the IRQ line for the currently active interrupt. See ["Figure 6 Interrupt life cycle - edge-triggered interrupts"](https://documentation-service.arm.com/static/5f1068720daa596235e7f6ef) in the GICv3 docs, which explains it better, but it also applies to GICv2.

> Is this the “special code for level-triggered IRQs'' you are referring to?

No, I meant the if-check code that uses `is_irq_edge_triggered()`.

> The GICv2 docs indicate that this shouldn’t affect the pending status for level-triggered interrupts as the kernel never writes to GICD_ISPENDR.

The main way interrupts become pending is from HW, not SW, so that's not relevant.

Level-triggered interrupts are always active and pending because the hardware still asserts the IRQ line when the interrupt becomes active. But because seL4 disables the IRQ till user-space ACKs the interrupt, it doesn't matter. For level-triggered IRQs user space will do something with the device to make the IRQ become idle again, so when it's done and it calls `seL4_IRQHandler_Ack()`, the pending will have disappeared (calling the ACK before handling the device would be a bug).

Edge-triggered interrupts behave a lot more like seL4 notifications. For those you need to call `seL4_IRQHandler_Ack()` first before handling the device, otherwise you may miss events and that would be a bug. I would even go as far as saying that you wouldn't need to call `seL4_IRQHandler_Ack()` for edge-triggered interrupts.

Common state transition for edge-triggered interrupts:

1. Edge -> IRQ pending
2. seL4 gets the `active_irq` by reading `int_ack` -> IRQ active
3. ...
4. seL4 calls `dist_pending_clr()` -> No change in IRQ state
5. seL4 Acknowledges the interrupt by writing EIO and signals user space -> IRQ idle
6. ...
7. User space calls `seL4_IRQHandler_Ack()` -> No change in IRQ state

Rare state transition for edge-triggered interrupts when another edge happens during kernel IRQ handling:

1. Edge -> IRQ pending
2. seL4 gets the `active_irq` by reading `int_ack` -> IRQ active
3. Edge -> IRQ active & pending
4. seL4 calls `dist_pending_clr()` -> IRQ active
5. seL4 Acknowledges the interrupt by writing EIO and signals user space -> IRQ idle
6. ...
7. User space calls `seL4_IRQHandler_Ack()` -> No change in IRQ state

So the only change clearing pending has is that edges happening while seL4 is in the interrupt handler are ignored. Any edged at 6 will still cause a new interrupt, even though user space was just notified an IRQ happed:

Less rare state transition for edge-triggered interrupts when another edge happens during user space IRQ handling:

1. Edge -> IRQ pending
2. seL4 gets the `active_irq` by reading `int_ack` -> IRQ active
3. (Nothing happens)
4. seL4 calls `dist_pending_clr()` -> No change in IRQ state
5. seL4 Acknowledges the interrupt by writing EIO and signals user space -> IRQ idle
6. Edge -> IRQ pending
7. User space calls `seL4_IRQHandler_Ack()` -> No change in IRQ state, but IRQ unmasked
8. seL4 gets the interrupt by reading `int_ack` -> IRQ active

Covering the race at 3, but not the bigger window of 6 doesn't make any sense. Either cover windows 6 and do something special during `seL4_IRQHandler_Ack()`, or remove all edge-triggered special casing. I propose the latter, because edge-triggered interrupts are not expected to happen often. Making all interrupt handling slightly faster and the code simpler is the better trade-off.

I'm going to open a new pull request to remove the edge-triggered special code, with the same explanation.